refactor: renaming and docs
