+++ sbi/neural_nets/density_estimators/__init__.py
@@ -3,8 +3,11 @@ from sbi.neural_nets.density_estimators.categorical_net import (
     CategoricalMassEstimator,
     CategoricalNet,
 )
+from sbi.neural_nets.density_estimators.flowmatching_estimator import (
+    FlowMatchingEstimator,
+)
 from sbi.neural_nets.density_estimators.mixed_density_estimator import (
     MixedDensityEstimator,
 )
 from sbi.neural_nets.density_estimators.nflows_flow import NFlowsFlow
-from sbi.neural_nets.density_estimators.zuko_flow import FlowMatchingEstimator, ZukoFlow
+from sbi.neural_nets.density_estimators.zuko_flow import ZukoFlow
